# server/docs/swagger.yaml
basePath: /
definitions:
...
  api.Error:
    type: object
...
  api.SocietyUserName:
    properties:
      id:
        example: 999
        type: integer
      insert_time:
        type: string
      last_update_time:
        type: string
      level:
        type: integer
      societyId:
        type: integer
      userId:
        type: integer
      userName:
        type: string
    required:
    - id
    - level
    - societyId
    - userId
    type: object
...
  model.SettingsPostMetadata:
    properties:
      name:
        type: string
      tooltip:
        type: string
      type:
        type: string
    type: object
  model.Society:
    properties:
      id:
        example: 999
        type: integer
      insert_time:
        type: string
      last_update_time:
        type: string
      loginURL:
        type: string
      name:
        type: string
      postMetadata:
        items:
          $ref: '#/definitions/model.SettingsPostMetadata'
        type: array
      secretKey:
        type: string
    required:
    - id
    - name
    - secretKey
    type: object
  model.SocietyIn:
    properties:
      loginURL:
        type: string
      name:
        type: string
      postMetadata:
        items:
          $ref: '#/definitions/model.SettingsPostMetadata'
        type: array
      secretKey:
        type: string
    required:
    - name
    - secretKey
    type: object
  model.SocietySummary:
    properties:
      id:
        example: 999
        type: integer
      name:
        type: string
    required:
    - id
    - name
    type: object
  model.SocietyUser:
    properties:
      id:
        example: 999
        type: integer
      insert_time:
        type: string
      last_update_time:
        type: string
      level:
        type: integer
      societyId:
        type: integer
      userId:
        type: integer
    required:
    - id
    - level
    - societyId
    - userId
    type: object
...
host: api.ourroots.org
info:
  contact:
    email: jim@anconafamily.com
    name: Jim Ancona
    url: https://github.com/jancona
  description: This is the OurRoots API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: OurRoots API
  version: 0.1.0
paths:
...
  /societies:
    post:
      consumes:
      - application/json
      operationId: addSociety
      parameters:
      - description: Add Society
        in: body
        name: society
        required: true
        schema:
          $ref: '#/definitions/model.SocietyIn'
      produces:
      - application/json
      responses:
        "201":
          description: OK
          schema:
            $ref: '#/definitions/model.Society'
        "415":
          description: Bad Content-Type
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: adds a new Society
      tags:
      - societies
  /societies/{id}:
    delete:
      operationId: deleteSociety
      parameters:
      - description: Society ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: OK
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: deletes a Society
      tags:
      - societies
    put:
      consumes:
      - application/json
      operationId: updateSociety
      parameters:
      - description: Society ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Society
        in: body
        name: society
        required: true
        schema:
          $ref: '#/definitions/model.Society'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Society'
        "415":
          description: Bad Content-Type
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: updates a Society
      tags:
      - societies
  /societies/{society}:
    get:
      operationId: getSociety
      parameters:
      - description: Society ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Society'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: gets a Society
      tags:
      - societies
  /societies/{society}/current_user:
    get:
      operationId: getCurrentSocietyUser
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SocietyUser'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: returns the current SocietyUser
      tags:
      - societyUsers
...
  /societies/{society}/users:
    get:
      operationId: getUsers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.SocietyUserName'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: returns all users
      tags:
      - societyUsers
  /societies/{society}/users/{id}:
    delete:
      operationId: deleteSocietyUSer
      parameters:
      - description: SocietyUser ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: OK
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: deletes a SocietyUser
      tags:
      - societyUsers
    put:
      consumes:
      - application/json
      operationId: updateSocietyUser
      parameters:
      - description: SocietyUser ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update SocietyUser
        in: body
        name: society
        required: true
        schema:
          $ref: '#/definitions/api.SocietyUserName'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SocietyUserName'
        "415":
          description: Bad Content-Type
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: updates a SocietyUser
      tags:
      - societyUsers
  /society_summaries:
    get:
      operationId: getSocietySummaries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.SocietySummary'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: returns all society summaries for the current user
      tags:
      - societies
  /society_summaries/{society}:
    get:
      operationId: getSocietySummary
      parameters:
      - description: Society ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SocietySummary'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/api.Error'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/api.Error'
      security:
      - OAuth2Implicit:
        - cms
        - openid
        - profile
        - email
      - OAuth2AuthCode:
        - cms
        - openid
        - profile
        - email
      summary: gets a SocietySummary
      tags:
      - societies
schemes:
- http
- https
securityDefinitions:
  OAuth2Implicit:
    authorizationUrl: https://ourroots.auth0.com/authorize?audience=https%3A%2F%2Fapi.ourroots.org%2Fpreprod
    flow: implicit
    scopes:
      cms: ' Grants read and write access to the CMS'
      email: ' Grants access to OIDC email attributes'
      openid: ' Indicates that the application intends to use OIDC to verify the user''s
        identity'
      profile: ' Grants access to OIDC user profile attributes'
    type: oauth2
swagger: "2.0"